Add MediaEngineEMEUWPSample
This commit adds a sample UWP application written in C++ which demonstrates protected Playready EME media playback using the MediaFoundation MediaEngine API and the WinRT composition APIs. This sample is an extended version of the MediaEngineUWPSample with the addition of protected playback support via the new CDM APIs added in 19041.